+++ changelog.rst
@@ -7,7 +7,7 @@ Bugfixes:
     - cosmic removal is now working again
     - fixes issue with auto peak number increasing while the checkbox for it was unchecked
     - no more error message when mouse is hovering over the cake image at 0 indices of the image
-    - changing the color of an overlay item is now working again
+    - changing the color of an overlay or phase item is now working correctly again
 
 
 0.5.7 (stable 24.04.2023)